fix/creacion componente modulo y quizz , metodo POST en el flux de modulos-componentAndflux.js
